refactor: split hrp.py into submodules and mirror test layout 1:1
Resolves two rhiza quality-gate findings.

#731 — Improve maintainability of src/pyhrp/hrp.py
Extract the covariance estimators (compute_cov/compute_corr/_returns) into
`pyhrp.covariance` and the Dendrogram dataclass plus its tree-building and
validation helpers (build_tree, _compute_distance_matrix, _bisect_tree,
_get_linkage, _check_finite_correlations, _validate_correlation_matrix,
_to_cluster) into `pyhrp.dendrogram`. `pyhrp.hrp` keeps only the hrp/schur_hrp
entry points and re-exports the moved public names, so the public API
(pyhrp.hrp and pyhrp) is unchanged. hrp.py maintainability index rises from
57.98 to 72.77 (and drops well under 250 LOC).

#740 — Restructure tests into a 1:1 mirrored layout
Consolidate the flat topical test files into tests/pyhrp/test_<module>.py
mirroring each source module, with TestPortfolio/TestCluster, TestNode and
TestDendrogram classes for the source classes, so
`scripts/check_test_layout.py` exits 0. All behavioral tests are preserved;
the redundant benchmark/stress smoke tests (whose CI workflows were removed in
#737) are dropped as they add no source coverage.

Gates: check_test_layout OK, ruff, mypy --strict + ty, deptry, bandit,
interrogate (100%), and the full suite (110 passed, 100% coverage) all green.
