+++ tests/pyhrp/__init__.py
@@ -1,0 +1,1 @@
+"""Test package mirroring the pyhrp source layout 1:1."""
