--- tests/test_hrp.py
@@ -1,130 +1,0 @@
-"""Tests for the hierarchical clustering and dendrogram building functionality."""
-
-from pathlib import Path
-
-import numpy as np
-import pytest
-from polars import DataFrame
-
-from pyhrp.hrp import _bisect_tree, _get_linkage, build_tree, compute_corr
-
-
-def test_linkage(returns: DataFrame, resource_dir: Path) -> None:
-    """Test the linkage matrix generation in the build_tree function.
-
-    This test verifies:
-    1. The correct order of node IDs in the dendrogram
-    2. The linkage matrix matches the expected values from a reference file
-
-    Args:
-        returns: DataFrame of asset returns
-        resource_dir: Path to test resources directory
-    """
-    dendrogram = build_tree(cor=compute_corr(returns), method="single", bisection=False)
-
-    assert dendrogram.ids == [11, 7, 19, 6, 14, 5, 10, 13, 3, 1, 4, 16, 0, 2, 17, 9, 8, 18, 12, 15]
-
-    np.testing.assert_array_almost_equal(dendrogram.linkage, np.loadtxt(resource_dir / "links.csv", delimiter=","))
-
-
-def test_bisection(returns: DataFrame, resource_dir: Path) -> None:
-    """Test the bisection method in the build_tree function.
-
-    This test verifies that the order of node IDs remains consistent
-    when using the bisection method.
-
-    Args:
-        returns: DataFrame of asset returns
-        resource_dir: Path to test resources directory
-    """
-    dendrogram = build_tree(cor=compute_corr(returns), method="single", bisection=True)
-
-    assert dendrogram.ids == [11, 7, 19, 6, 14, 5, 10, 13, 3, 1, 4, 16, 0, 2, 17, 9, 8, 18, 12, 15]
-
-
-def test_plot_bisection(returns: DataFrame) -> None:
-    """Test building a dendrogram with bisection and verify the node order.
-
-    This test verifies:
-    1. The dendrogram is properly constructed with bisection
-    2. The dendrogram has the expected structure (root and linkage)
-    3. The order of asset names in the dendrogram matches the expected order
-
-    Args:
-        returns: DataFrame of asset returns
-    """
-    cor = compute_corr(returns)
-    dendrogram = build_tree(cor=cor, method="single", bisection=True)
-
-    assert dendrogram.root is not None
-    assert dendrogram.linkage is not None
-
-    assert dendrogram.names == [
-        "UAA",
-        "WMT",
-        "SBUX",
-        "AMD",
-        "RRC",
-        "GE",
-        "T",
-        "XOM",
-        "BABA",
-        "AAPL",
-        "AMZN",
-        "MA",
-        "GOOG",
-        "FB",
-        "PFE",
-        "GM",
-        "BAC",
-        "JPM",
-        "SHLD",
-        "BBY",
-    ]
-
-
-@pytest.mark.parametrize("method", ["single", "ward", "average", "complete"])
-def test_invariant_order(returns: DataFrame, method: str) -> None:
-    """Test that the order of nodes is invariant to the bisection parameter.
-
-    This test verifies that regardless of whether bisection is used or not,
-    the resulting dendrogram maintains the same order of assets, IDs, and names
-    for different clustering methods.
-
-    Args:
-        returns: DataFrame of asset returns
-        method: Clustering method to use (single, ward, average, or complete)
-    """
-    cor = compute_corr(returns)
-    dendrogram1 = build_tree(cor=cor, method=method, bisection=True)
-    dendrogram2 = build_tree(cor=cor, method=method, bisection=False)
-
-    assert dendrogram1.assets == dendrogram2.assets
-    assert dendrogram1.ids == dendrogram2.ids
-    assert dendrogram1.names == dendrogram2.names
-
-
-def test_bisect_tree_helper() -> None:
-    """Test the module-level _bisect_tree helper."""
-    root, next_id = _bisect_tree(ids=[0, 1, 2, 3], next_id=3)
-
-    assert next_id == 6
-    assert root.value == 6
-    assert [leaf.value for leaf in root.leaves] == [0, 1, 2, 3]
-
-
-def test_bisect_tree_helper_empty_ids() -> None:
-    """Test _bisect_tree rejects empty ids input."""
-    with pytest.raises(ValueError, match="at least one node id"):
-        _bisect_tree(ids=[], next_id=0)
-
-
-def test_get_linkage_helper() -> None:
-    """Test the module-level _get_linkage helper."""
-    root, _ = _bisect_tree(ids=[0, 1, 2, 3], next_id=3)
-
-    assert _get_linkage(root) == [
-        [0.0, 1.0, 3.0, 2.0],
-        [2.0, 3.0, 3.0, 2.0],
-        [4.0, 5.0, 7.0, 4.0],
-    ]
